requirements.txt
asgiref==3.3.1
Django==3.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-redis==4.12.1
mysqlclient==2.0.3
Pillow==8.1.2
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
